Refatoracao no codigo
