--- src/App.js
@@ -1,55 +1,0 @@
-import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import Home from './routes/Home';
-import Login from './routes/Login';
-import Cadastro from './routes/Cadastro';
-import Comunidades from './routes/Comunidades';
-import Footer from './components/Footer';
-import { auth } from './firebase-config';
-import { Navigate } from 'react-router-dom';
-import { useState, useEffect } from 'react';
-
-
-function App() {
-  const [loading, setLoading] = useState(true);
-  const [isAuthenticated, setIsAuthenticated] = useState(false);
-
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      if (user) {
-        setIsAuthenticated(true);
-      } else {
-        setIsAuthenticated(false);
-      }
-      setLoading(false);
-    });
-
-    return () => unsubscribe();
-  }, []);
-
-  if (loading) {
-    return <div>Carregando...</div>;
-  }
-
-  return (
-    <Router>
-      <div className="flex flex-col min-h-screen">
-        <main className="flex-grow flex flex-col">
-          <Routes>
-            <Route path="/" element={<Home />} />
-            <Route path="/login" element={<Login />} />
-            <Route path="/cadastro" element={<Cadastro />} />
-            <Route
-              path="/comunidades"
-              element={
-                auth.currentUser ? <Comunidades /> : <Navigate to="/login" />
-              }
-            />
-          </Routes>
-        </main>
-        <Footer />
-      </div>
-    </Router>
-  );
-}
-
-export default App;
